Introduce the LiquidLib concept
This commit tries a different design to store the spec into interfaces
(as annotations). Instead of bundling everything up into a single, fat
spec, this commit tries to store the spec plus any dependency-spec that
was necessary to refine that module.

Last but not least we now look for cached specs into the interfaces all
the time; for home modules we use the `HomePackageTable`. This allows us
to get rid of the SpecEnv abstraction altogether.